+++ cmd/kyma/alpha/deploy/cmd.go
@@ -3,6 +3,7 @@ package deploy
 import (
 	"context"
 	"fmt"
+	"github.com/kyma-project/cli/internal/coredns"
 	"io/fs"
 	"io/ioutil"
 	"strings"
@@ -139,7 +140,7 @@ func (cmd *command) Run(o *Options) error {
 		return err
 	}
 
-	if _, err := overrides.PatchCoreDNS(cmd.K8s.Static(), overridesBuilder, isK3d); err != nil {
+	if _, err := coredns.PatchCoreDNS(cmd.K8s.Static(), overridesBuilder, isK3d); err != nil {
 		return err
 	}
